+++ cloud/aws/kubeconfig.go
@@ -2,12 +2,18 @@ package aws
 
 import (
 	"fmt"
+	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"regexp"
 	"sync"
 
 	log "github.com/Sirupsen/logrus"
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/aws/aws-sdk-go/service/s3"
+	"github.com/aws/aws-sdk-go/service/s3/s3manager"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
@@ -18,12 +24,13 @@ var cacheClientsetOnce sync.Once
 var cacheConfigPathOnce sync.Once
 
 var (
-	rootPath      = "/opt/datacol"
-	kcpath        = filepath.Join(rootPath, "kubeconfig")
-	pemPathRE     = filepath.Join(rootPath, "%s.pem")
-	privateIpAttr = "MasterPrivateIp"
-	bastionIpAttr = "BastionHostPublicIp"
-	scpCmd        = "scp -i %s -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null ubuntu@%s:~/kubeconfig %s"
+	rootPath            = "/opt/datacol"
+	kcpath              = filepath.Join(rootPath, "kubeconfig")
+	pemPathRE           = filepath.Join(rootPath, "%s.pem")
+	privateIpAttr       = "MasterPrivateIp"
+	bastionIpAttr       = "BastionHostPublicIp"
+	scpCmd              = "scp -i %s -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null ubuntu@%s:~/kubeconfig %s"
+	s3StateStoreMatcher = regexp.MustCompile(`s3://(\S+)/cluster-info.yaml`)
 )
 
 func (p *AwsCloud) kubeClient() *kubernetes.Clientset {
@@ -43,9 +50,46 @@ func (p *AwsCloud) kubeClient() *kubernetes.Clientset {
 	return kubeClient
 }
 
+// fetchClusterInfoFromS3 tries to fetch cluster state from s3. Currently not being used
+// since the cluster-info.yaml doesn't contain user credentials
+func (p *AwsCloud) fetchClusterInfoFromS3(path string) (cached bool, err error) {
+	value, err := p.stackOutputValue("JoinNodes")
+	log.Warn(err)
+
+	if value != "" {
+		log.Debugf("Found JoinNodes Output: %s", value)
+
+		if match := s3StateStoreMatcher.FindStringSubmatch(value); len(match) == 2 {
+			s3bucket := match[1]
+			buff := &aws.WriteAtBuffer{}
+			s3dl := s3manager.NewDownloader(session.New())
+
+			if _, err = s3dl.Download(buff, &s3.GetObjectInput{
+				Bucket: aws.String(s3bucket),
+				Key:    aws.String("cluster-info.yaml"),
+			}); err != nil {
+				return
+			}
+
+			err = ioutil.WriteFile(path, buff.Bytes(), 0644)
+			if err == nil {
+				cached = true
+			}
+		}
+	}
+
+	return
+}
+
 func (p *AwsCloud) K8sConfigPath() (string, error) {
 	if _, err := os.Stat(kcpath); err != nil {
 		if os.IsNotExist(err) {
+
+			// if fetched, err := p.fetchClusterInfoFromS3(kcpath); fetched {
+			// 	log.Infof("cached the %s from s3 state store. err:%v", kcpath, err)
+			// 	return kcpath, nil
+			// }
+
 			ipAddr, err := p.masterPrivateIp()
 			if err != nil {
 				return ipAddr, err