+++ cmd/apps.go
@@ -3,6 +3,9 @@ package cmd
 import (
 	"fmt"
 	"os"
+	"strings"
+
+	"github.com/olekukonko/tablewriter"
 
 	log "github.com/Sirupsen/logrus"
 	term "github.com/appscode/go/term"
@@ -16,6 +19,7 @@ func init() {
 		Name:   "apps",
 		Usage:  "Manage your apps in a stack",
 		Action: cmdAppsList,
+		Flags:  []cli.Flag{&stackFlag},
 		Subcommands: []cli.Command{
 			cli.Command{
 				Name:   "create",
@@ -75,7 +79,14 @@ func cmdAppsList(c *cli.Context) error {
 	if len(apps) == 0 {
 		fmt.Println("No apps found.")
 	} else {
-		fmt.Println(toJson(apps))
+		table := tablewriter.NewWriter(os.Stdout)
+		table.SetHeader([]string{"NAME", "BUILD", "RELEASE", "DOMAINS"})
+		for _, a := range apps {
+			table.Append([]string{a.Name, a.BuildId, a.ReleaseId,
+				strings.Join(a.Domains, "\n"),
+			})
+		}
+		table.Render()
 	}
 	return nil
 }