+++ cmd/build.go
@@ -18,6 +18,7 @@ import (
 	"github.com/docker/docker/builder/dockerignore"
 	"github.com/docker/docker/pkg/archive"
 	"github.com/docker/docker/pkg/fileutils"
+	"github.com/olekukonko/tablewriter"
 	"github.com/urfave/cli"
 	"golang.org/x/net/context"
 	"golang.org/x/net/websocket"
@@ -39,6 +40,13 @@ func init() {
 				Name:   "list",
 				Usage:  "get builds for an app",
 				Action: cmdBuildList,
+				Flags: []cli.Flag{
+					&cli.IntFlag{
+						Name:  "limit, n",
+						Usage: "Limit the number of recent builds to fetch",
+						Value: 5,
+					},
+				},
 			},
 			{
 				Name:   "delete",
@@ -57,10 +65,17 @@ func cmdBuildList(c *cli.Context) error {
 	api, close := getApiClient(c)
 	defer close()
 
-	builds, err := api.GetBuilds(name)
+	builds, err := api.GetBuilds(name, c.Int("limit"))
 	stdcli.ExitOnError(err)
 
-	fmt.Println(toJson(builds))
+	table := tablewriter.NewWriter(os.Stdout)
+	table.SetHeader([]string{"ID", "COMMIT", "STATUS", "CREATED"})
+	for _, b := range builds {
+		delta := elaspedDuration(time.Unix(int64(b.CreatedAt), 0))
+		table.Append([]string{b.Id, b.Version, b.Status, delta})
+	}
+
+	table.Render()
 	return nil
 }