+++ cmd/provider/aws/templates.go
@@ -83,7 +83,7 @@ func cmdProviderAwsTemplatesFormationYaml() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "cmd/provider/aws/templates/formation.yaml", size: 22170, mode: os.FileMode(436), modTime: time.Unix(1521183917, 0)}
+	info := bindataFileInfo{name: "cmd/provider/aws/templates/formation.yaml", size: 22170, mode: os.FileMode(420), modTime: time.Unix(1521208860, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }